缓存逻辑优化，确保每一级的id都是唯一的

+++ library/src/main/java/com/zr/addressselector/model/City.java
@@ -1,7 +1,7 @@
 package com.zr.addressselector.model;
 
 public class City  {
-    public int id;
-    public int province_id;
+    public long id;
+    public long province_id;
     public String name;
 }
