+++ library/src/main/java/com/zr/addressselector/model/County.java
@@ -1,7 +1,7 @@
 package com.zr.addressselector.model;
 
 public class County {
-    public int id;
-    public int city_id;
+    public long id;
+    public long city_id;
     public String name;
 }
