+++ library/src/main/java/com/zr/addressselector/model/Street.java
@@ -1,7 +1,7 @@
 package com.zr.addressselector.model;
 
 public class Street {
-    public int id;
-    public int county_id;
+    public long id;
+    public long county_id;
     public String name;
 }
